Agregando los identificadores de tipo long.

+++ src/ve/gob/cnti/srsi/modelo/Correo.java
@@ -1,13 +1,5 @@
 package ve.gob.cnti.srsi.modelo;
 
-/****
- *
- * @author joaquin
- * 
- * Clase modelo con los atributos de la tabla correos
- *
- */
-
 import java.util.Date;
 
 import javax.persistence.Entity;
@@ -16,69 +8,90 @@ import javax.persistence.GenerationType;
 import javax.persistence.Id;
 import javax.persistence.Table;
 
+/**
+ * Clase modelo con los atributos de la tabla correos.
+ * 
+ * @author Joaquín Pereira
+ * 
+ */
 @Entity
-@Table(name="correos")
+@Table(name = "correos")
 public class Correo {
-	
+
 	@Id
-	@GeneratedValue(strategy=GenerationType.IDENTITY)
-	private int id;
-	private int id_correo;
-	private int id_padre;
-	private int origen;
+	@GeneratedValue(strategy = GenerationType.IDENTITY)
+	private long id;
+	private long id_correo;
+	private long id_padre;
+	private long origen;
 	private String correo;
 	private int status;
 	private Date fecha_creado;
 	private Date fecha_modificado;
-	
-	public Correo() {	
+
+	public Correo() {
 	}
-	
-	public int getId() {
+
+	public long getId() {
 		return id;
 	}
-	public void setId(int id) {
+
+	public void setId(long id) {
 		this.id = id;
 	}
-	public int getId_correo() {
+
+	public long getId_correo() {
 		return id_correo;
 	}
-	public void setId_correo(int id_correo) {
+
+	public void setId_correo(long id_correo) {
 		this.id_correo = id_correo;
 	}
-	public int getId_padre() {
+
+	public long getId_padre() {
 		return id_padre;
 	}
-	public void setId_padre(int id_padre) {
+
+	public void setId_padre(long id_padre) {
 		this.id_padre = id_padre;
 	}
-	public int getOrigen() {
+
+	public long getOrigen() {
 		return origen;
 	}
-	public void setOrigen(int origen) {
+
+	public void setOrigen(long origen) {
 		this.origen = origen;
 	}
+
 	public String getCorreo() {
 		return correo;
 	}
+
 	public void setCorreo(String correo) {
 		this.correo = correo;
 	}
+
 	public int getStatus() {
 		return status;
 	}
+
 	public void setStatus(int status) {
 		this.status = status;
 	}
+
 	public Date getFecha_creado() {
 		return fecha_creado;
 	}
+
 	public void setFecha_creado(Date fecha_creado) {
 		this.fecha_creado = fecha_creado;
 	}
+
 	public Date getFecha_modificado() {
 		return fecha_modificado;
 	}
+
 	public void setFecha_modificado(Date fecha_modificado) {
 		this.fecha_modificado = fecha_modificado;
 	}
@@ -90,5 +103,5 @@ public class Correo {
 				+ ", status=" + status + ", fecha_creado=" + fecha_creado
 				+ ", fecha_modificado=" + fecha_modificado + "]";
 	}
-	
+
 }
